Add "check" operation
In God mode, this is a successful no-op (Allah doesn't have a
"self-test".)

In Allah mode, this runs a check command for each task, if it's up. If
the check command fails, the task is deemed to have "failed".

I'm keeping "status" and "check" separate so that an Allah resource
agent can use them respectively in its "status" and "monitor" actions.