Don't look for svscan in procfs
Looking for an executable names .../svscan is fragile: on Linux, if the
executable has been replaced (say, by an upgrade or by prelink), the exe
entry will be suffixed with " (deleted)". It may also erroneously match
on some other executable that just happens to be named svscan.

Instead, simply assume services are configured in /etc/service/. This
can be overridden with the ALLAH_SERVICE_DIR environment variable.